summary: google, github, magic link and email-password working, db with supabase and prisma working, email server in Brevo active with mails from personal gmail for now
feat(auth): enhance Google authentication options
Adds additional parameters to the Google authentication 
configuration to request user consent and enable offline 
access. This improves the user experience by ensuring 
that the application can access user data even when 
the user is not actively using the app. 

Removes the shadow database URL from the Prisma schema 
as it is no longer necessary for the current setup. 

Updates the .gitignore file to include the production 
environment variables file to prevent it from being 
tracked in version control.